+++ CacheWarmer/DependencyCacheWarmer.php
@@ -53,7 +53,7 @@ class DependencyCacheWarmer implements CacheWarmerInterface
      */
     public function isOptional()
     {
-        return false;
+        return true;
     }
 
     /**